Add create and get operations
The 'create' operations creates a pdf file with the spec received, saves it
in a temporary directory and returns the URL of the file. The file can be
retrieved using the 'get' operation.

Problems have been found with the permissions on the temporary directory with
the server returning an 'Access is denied' error. Giving full control to the
NETWORK SERVICE user has not fixed it, so we need to investigate a little bit
on this problem.

For the 'get' operation, MapFish uses the FileApp method from Paste, which
handles all the things related with serving files. A we can not use it, for now
we will only handle 404 and 403 errors.