Add a configuration file
An INI file containing configuration options has been added. The file is parsed
with SafeConfigParse and contains options for defining the JAR file path, the
YAML file path and the temporal directory. If the temporary directory is not
defined, the OS temp dir is used (i.e. the one that returns gettempdir())