Add a default handler for non print related requests
All requests that don't match one of the printing service endpoints are
forwarded to a default handler. This allows using the same installation
of PyISAPIe across multiple applications. E.g. you can have a virtual
directory called "app" with PyISAPIe configured as "Script Map" and
another one called "printing" with PyISAPIe configured as "Wildcard
Script Map".

The default handler is a slightly modified version of the default
Isapi.py file that returns a 404 error if the requested file is not
found.

As the PyISAPIe author points out in the default Isapi.py file, is
always better to define a mapping handler rather than forward all
requests:
"I don't really recommend this method - it's not as
package/module oriented as the Regex and Advanced
examples, which can handle arbitrary URLs and pass
them to preloaded handlers."